Don't return a nil thumbnail path!
Is there any situation where we'd rather have a nil return than the
default image?  Given that the only time this block hits is when
ActiveFedora fails to find something, I'm really not sure... but that
should be a very exceptional case, and it gets logged as an error
already, so....  Just gonna avoid returning nil.